+++ src/components/NotFoundPage.tsx
@@ -1,0 +1,3 @@
+export const NotFoundPage = () => (
+  <h1 className="title">Page not found</h1>
+);
