+++ src/components/PeoplePage.tsx
@@ -12,6 +12,7 @@ export const PeoplePage = () => {
 
   useEffect(() => {
     setLoading(true);
+    setErrorMessage('');
 
     getPeople()
       .then(data => {
