+++ src/components/PeopleTable.tsx
@@ -1,6 +1,16 @@
-import { PersonPage } from './PersonPage';
 
-export const PeopleTable = ({ people }) => {
+import { Person } from '../types/Person';
+import { PersonLink } from './PersonLink';
+import { useParams } from 'react-router-dom';
+
+
+interface Props {
+  people: Person[];
+}
+
+export const PeopleTable: React.FC<Props> = ({ people }) => {
+  const { slug } = useParams();
+
   return (
     <table
       data-cy="peopleTable"
@@ -18,9 +28,36 @@ export const PeopleTable = ({ people }) => {
       </thead>
 
       <tbody>
-        {people.map(person => (
-          <PersonPage key={person.slug} person={person} />
-        ))}
+        {people.map(person => {
+          const isSelected = person.slug === slug;
+
+          return (
+            <tr
+              key={person.slug}
+              data-cy="person"
+              className={isSelected ? 'has-background-warning' : ''}
+            >
+              <td>
+                <PersonLink person={person} />
+              </td>
+              <td>{person.sex}</td>
+              <td>{person.born}</td>
+              <td>{person.died}</td>
+              <td>
+                <PersonLink
+    person={person.mother}
+    name={person.motherName}
+  />
+              </td>
+              <td>
+                <PersonLink
+    person={person.father}
+    name={person.fatherName}
+  />
+              </td>
+            </tr>
+          );
+        })}
       </tbody>
     </table>
   );
